Moved tests to seperate directory.
